make a little changes on fullPage to make it swipe
Make fullPage script more like swiper when it scrolls horizontally.
The default swipe behavior is more like 1 2 3 and it scrolls back to
the first one backwards.

To make it behaves like a smooth frontwards, it requires 2 caches: one
is the copy of the first one but appended to be last, another is the
copy of original last one but prepended to the first.

Suppose we have 3 pages originally, the html page should duplicate the
first slide and last slide to last and first, set options.
continuousVerticalLoop true, when it swipes horizontally from 3 to 1,
the new script would use the duplicated one at the last, when user
scrolls from 1 to 2, in the mean time just before it scrolls, script
would silent landscape to real 1 in no time, and then execute an
animated 1 to 2. For 1 to 3 vise versa.